% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnll.R
\name{fnll}
\alias{fnll}
\title{Fonction RTMB à optimiser.}
\usage{
fnll()
}
\value{
La valeur de vraissemblance calculée
}
\description{
Les objets dd_data et dd_param doivent exister au moment où ce modèle est optimisé.
}
\examples{
#À venir
}
