% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dd.R
\name{run_dd}
\alias{run_dd}
\title{Exécute l'ajustement du modèle Delay-Difference en RTMB}
\usage{
run_dd()
}
\value{
Null
}
\description{
Exécute l'ajustement du modèle Delay-Difference en RTMB
}
\examples{
run_dd()
}
